add FileSystemOperator class

+++ client/client_daemon.py
@@ -20,13 +20,13 @@ def req_post(*args, **kwargs):
 	try:
 		return requests.post(*args, **kwargs)
 	except requests.exceptions.RequestException:
-		return False # TODO True x test | use False
+		return True # TODO True x test | use False
 
 def req_get(*args, **kwargs):
 	try:
 		return requests.get(*args, **kwargs)
 	except requests.exceptions.RequestException:
-		return True # TODO True x test | use False
+		return False # TODO True x test | use False
 
 def req_put(*args, **kwargs):
 	try:
@@ -61,6 +61,9 @@ class ServerCommunicator(object):
 		print success
 		return request_result
 
+	def get_fullpath(self, dst_path):
+		return os.path.join(self.dir_path, dst_path)
+
 	def download_file(self, dst_path):
 		""" download a file from server"""
 
@@ -73,43 +76,28 @@ class ServerCommunicator(object):
 			"url": server_url,
 			"auth": self.auth
 		}
-		print request
 		r = self._try_request(req_get, success_log, error_log, **request)
-		
-		local_path = os.path.join(self.dir_path, dst_path)
-
-		#create a directory chain without file name
-		os.makedirs(local_path[-1], 0755 )
-		#create a downloaded file
-		with open(local_path, 'w') as f:
-			f.write(r.text)
+		local_path = self.get_fullpath(dst_path)
+		return local_path, r.text
+	
 
 	def upload_file(self, dst_path, put_file = False):
 		""" upload a file to server """
 
-		file_name = dst_path.split(os.path.sep)[-1]
-		path = os.path.join(*(dst_path.split(os.path.sep)[0:-1]))
+		path = os.path.join(*(dst_path.split(os.path.sep)))
 		file_content = ''
-		
 		try:
-			with open(dst_path, 'r') as f:
-				file_content =  f.read()
+			file_content = open(dst_path, 'rb')
 		except IOError:
 			return False #Atomic create and delete error!
-
-		upload = {
-					'file_name': file_name,
-					'file_content': file_content
-				 }
 
 		server_url = "{}/files/{}".format(self.server_url, path)
 
 		error_log = "ERROR upload request " + dst_path
 		success_log = "file uploaded! " + dst_path
-		
 		request = {
 			"url": server_url,
-			"data": upload,
+			"files": {'file_content':file_content},
 			"auth": self.auth
 		}
 		if put_file:
@@ -138,7 +126,7 @@ class ServerCommunicator(object):
 		error_log = "User creation error"
 		success_log = "user created!" 
 
-		server_url = "{}/create_user".format(self.server_url)
+		server_url = "{}/user/create".format(self.server_url)
 
 		request = {
 			"url": server_url,
@@ -147,8 +135,88 @@ class ServerCommunicator(object):
 					"psw": password
 			}
 		}
-		
 		self._try_request(req_post, success_log, error_log, **request)
+
+class FileSystemOperator(object):
+	
+	def __init__(self, event_handler, server_com):
+		self.event_handler = event_handler
+		self.server_com = server_com
+
+	def send_lock(self, path):
+		self.event_handler.path_ignored = path
+
+	def send_unlock(self):
+		self.event_handler.path_ignored = None
+
+	def write_a_file(self, path):
+		"""
+		write a file (download if exist or not [get and put])
+
+			send lock to watchdog
+			download the file from server
+			create directory chain
+			create file
+			send unlock to watchdog
+		"""
+		full_path = self.server_com.get_fullpath(path)
+		self.send_lock(full_path)
+		fullpath, content = self.server_com.download_file(path)
+		try:
+			os.makedirs(os.path.split(fullpath)[0], 0755 )
+		except OSError:
+			pass
+		with open(fullpath, 'w') as f:
+			f.write(content)
+		time.sleep(3)
+		self.send_unlock()
+
+	def move_a_file(self, origin_path, dst_path):
+		""" 
+		move a file 
+
+			send lock to watchdog
+			create directory chain for dst_path
+			move the file from origin_path to dst_path
+			send unlock to watchdog
+		"""
+		self.send_lock(origin_path)
+		try:
+			os.makedirs(os.path.split(dst_path)[0], 0755 )
+		except OSError:
+			pass
+		shutil.move(origin_path, dst_path)
+		self.send_unlock()
+
+	def copy_a_file(self, origin_path, dst_path):
+		""" copy a file 
+
+			send lock to watchdog
+			create directory chain for dst_path
+			copy the file from origin_path to dst_path
+			send unlock to watchdog
+		"""
+		self.send_lock(origin_path)
+		try:
+			os.makedirs(os.path.split(dst_path)[0], 0755 )
+		except OSError:
+			pass
+		shutil.copyfile(origin_path, dst_path)
+		self.send_unlock()
+
+	def delete_a_file(self, path):
+		""" 
+		delete a file 
+			send lock to watchdog
+			delete file
+			send unlock to watchdog
+		"""
+		self.send_lock(origin_path)
+		try:
+			shutil.rmtree(path)
+		except IOError:
+			pass
+		self.send_unlock
 
 def load_config():
 	with open('config.json', 'r') as config_file:
@@ -160,10 +228,8 @@ class DirectoryEventHandler(FileSystemEventHandler):
 
 	def __init__(self, cmd):
 		self.cmd = cmd
+		self.path_ignored = None
 		
-	# def on_any_event(self, event):
-		# request = requests.get(self.server_url)
-
 	def on_moved(self, event):
 		"""Called when a file or a directory is moved or renamed.
 
@@ -172,8 +238,11 @@ class DirectoryEventHandler(FileSystemEventHandler):
 		:type event:
 			:class:`DirMovedEvent` or :class:`FileMovedEvent`
 		"""
-		self.cmd.delete_file(event.src_path)
-		self.cmd.upload_file(event.dest_path)
+		if self.path_ignored != event.src_path:
+			self.cmd.delete_file(event.src_path)
+			self.cmd.upload_file(event.dest_path)
+		else:
+			print "ingnored move on ", event.src_path
 
 	def on_created(self, event):
 		"""Called when a file or directory is created.
@@ -183,7 +252,10 @@ class DirectoryEventHandler(FileSystemEventHandler):
 		:type event:
 			:class:`DirCreatedEvent` or :class:`FileCreatedEvent`
 		"""
-		self.cmd.upload_file(event.src_path)
+		if self.path_ignored != event.src_path:
+			self.cmd.upload_file(event.src_path)
+		else:
+			print "ingnored create on ", event.src_path
 
 	def on_deleted(self, event):
 		"""Called when a file or directory is deleted.
@@ -193,7 +265,10 @@ class DirectoryEventHandler(FileSystemEventHandler):
 		:type event:
 			:class:`DirDeletedEvent` or :class:`FileDeletedEvent`
 		"""
-		self.cmd.delete_file(event.src_path)
+		if self.path_ignored != event.src_path:
+			self.cmd.delete_file(event.src_path)
+		else:
+			print "ingnored deletion on ", event.src_path
 
 	def on_modified(self, event):
 		"""Called when a file or directory is modified.
@@ -203,8 +278,12 @@ class DirectoryEventHandler(FileSystemEventHandler):
 		:type event:
 			:class:`DirModifiedEvent` or :class:`FileModifiedEvent`
 		"""
-		if not event.is_directory:
-			self.cmd.upload_file(event.src_path, put_file = True)
+		print self.path_ignored,event.src_path
+		if self.path_ignored != event.src_path:
+			if not event.is_directory:
+				self.cmd.upload_file(event.src_path, put_file = True)
+		else:
+			print "ingnored modified on ", event.src_path
 
 
 
@@ -218,13 +297,16 @@ def main():
 		password = config['password'],
 		dir_path = config['dir_path'])
 
-	#server_com.create_user("usernameFarlocco", "passwordSegretissima")
-	server_com.download_file('bla.txt')
 
 	event_handler = DirectoryEventHandler(server_com)
+	file_system_op = FileSystemOperator(event_handler, server_com)
 	observer = Observer()
 	observer.schedule(event_handler, config['dir_path'], recursive=True)
+
 	observer.start()
+	#server_com.create_user("usernameFarlocco", "passwordSegretissima")
+	#file_system_op.write_a_file('pippo/pippo/pluto/bla.txt') #test
+	#server_com.upload_file("/home/user/prove/bho/ciao.js")
 
 	try:
 		while True:
